Add electronics and eagle
